refactor(netsim): move most code inside netsim/netstack
The top-level package should contain more specific code handling
scenarios rather than the netstack stub network.

Let's make this possibly by moving the netstack into its own package.
