+++ netsim/netstack/errno_windows.go
@@ -6,7 +6,7 @@
 // Windows errno definitions.
 //
 
-package netsim
+package netstack
 
 import "golang.org/x/sys/windows"
